recommender/requirements.txt
fastapi==0.115.12
numpy==2.2.5
pandas==2.2.3
psycopg[binary]==3.2.6
pydantic==2.11.4
python-dotenv==1.1.0
requests==2.32.3
sentence-transformers==4.1.0
uvicorn==0.34.2
